routing fix and orders handling no user selected/or records found

+++ php_test/view/listOrder.php
@@ -72,44 +72,48 @@
                     </div>
                     <br>
                     <?php
-                        if($result->num_rows > 0){
+                    if(!empty($result)) {
+                        if ($result->num_rows > 0) {
                             echo "<table class='table table-bordered table-striped'>";
-                                echo "<thead>";
-                                    echo "<tr>";
-                                        echo "<th>Order#</th>";
-                                        echo "<th>Customer#</th>";
-                                        echo "<th>Purchase Date</th>";
-                                        echo "<th>Country</th>";
-                                        echo "<th>Device</th>";
-                                        echo "<th>EAN</th>";
-                                        echo "<th>Quantity</th>";
-                                        echo "<th>Price</th>";
-                                    echo "</tr>";
-                                echo "</thead>";
-                                echo "<tbody>";
-                                while($row = mysqli_fetch_array($result)){
-                                    echo "<tr>";
-                                        echo "<td>" . $row['id'] . "</td>";
-                                        echo "<td>" . $row['cId'] . "</td>";
-                                        echo "<td>" . $row['purchase_Date'] . "</td>";
-                                        echo "<td>" . $row['country'] . "</td>";
-                                        echo "<td>" . $row['device'] . "</td>";
-                                        echo "<td>" . $row['EAN'] . "</td>";
-                                        echo "<td>" . $row['quantity'] . "</td>";
-                                        echo "<td>" . $row['price'] . "</td>";
-                                        echo "<td>";
-                                        echo "<a href='index.php?act=update&id=". $row['id'] ."' title='Update Record' data-toggle='tooltip'><i class='fa fa-edit'></i></a>";
-                                        echo "<a href='index.php?act=delete&id=". $row['id'] ."' title='Delete Record' data-toggle='tooltip'><i class='fa fa-trash'></i></a>";
-                                        echo "</td>";
-                                    echo "</tr>";
-                                }
-                                echo "</tbody>";
+                            echo "<thead>";
+                            echo "<tr>";
+                            echo "<th>Order#</th>";
+                            echo "<th>Customer#</th>";
+                            echo "<th>Purchase Date</th>";
+                            echo "<th>Country</th>";
+                            echo "<th>Device</th>";
+                            echo "<th>EAN</th>";
+                            echo "<th>Quantity</th>";
+                            echo "<th>Price</th>";
+                            echo "</tr>";
+                            echo "</thead>";
+                            echo "<tbody>";
+                            while ($row = mysqli_fetch_array($result)) {
+                                echo "<tr>";
+                                echo "<td>" . $row['id'] . "</td>";
+                                echo "<td>" . $row['cId'] . "</td>";
+                                echo "<td>" . $row['purchase_Date'] . "</td>";
+                                echo "<td>" . $row['country'] . "</td>";
+                                echo "<td>" . $row['device'] . "</td>";
+                                echo "<td>" . $row['EAN'] . "</td>";
+                                echo "<td>" . $row['quantity'] . "</td>";
+                                echo "<td>" . $row['price'] . "</td>";
+                                echo "<td>";
+                                echo "<a href='index.php?act=update&id=" . $row['id'] . "' title='Update Record' data-toggle='tooltip'><i class='fa fa-edit'></i></a>";
+                                echo "<a href='index.php?act=delete&id=" . $row['id'] . "' title='Delete Record' data-toggle='tooltip'><i class='fa fa-trash'></i></a>";
+                                echo "</td>";
+                                echo "</tr>";
+                            }
+                            echo "</tbody>";
                             echo "</table>";
                             // Free result set
                             mysqli_free_result($result);
                         } else {
                             echo "<p class='lead'><em>No records were found.</em></p>";
                         }
+                    } else {
+                        echo "<p class='lead'><em>Please select 'See Orders' from customer table to display order records</em></p>";
+                    }
                     ?>
                 </div>
                 <div class="page-header clearfix">
